New entry in ServerStartup enum to indicate AddrInUse Error (#840)
when cargo fires of parallel build jobs, we encounter a race to start
the server if the server is not already running to start the server

This is an AddrInUse error causing atleast one of the parallel build
jobs fail.

This patch adds a new entry to ServerStartup enum to clearly communicate
the AddrInUse error from the server to the client that requested the
server start. Using this information the client can retry connecting
instead of failing.