msvc: handle /MP as a passthrough argument
Compiler invocations with the `/MP[processMax]` option for MSVC are
currently rejected by the cache ("multiple input files"). The `/MP`
option enables multiprocessing in the compiler [1], so there is no
impact on the cache-ability of the call if a single source file is
given. This commit changes the behavior to passthrough.

I ran into this because Conan [2] adds the `/MP` option by default.

[1]: https://docs.microsoft.com/en-us/cpp/build/reference/mp-build-with-multiple-processes
[2]: https://conan.io/