Move tests from Travis/Appveyor to Github Actions
On Travis, we used to build and run tests with both
--no-default-features and without, but without additional features, which
was the majority of cases, testing with --no-default-features was not
really useful, so we don't do that anymore, except for the dist-server.

Also, since dist-tests enables dist-server, the dist-server cases ended
up running the same tests multiple times, which we now avoid.

We also now split building the tests from running them, such that logs
can be a little easier to work with.

Github actions don't have arm64 runners yet, so we now don't run tests on
arm64 linux, but that's not expected to catch problems that wouldn't be
caught by the other tests.

The Appveyor setup was running tests with multiple versions of rust,
which is not done here, but that's not expected to catch problems that
wouldn't be caught by the other tests either.