C/C++: add extra input files via environment SCCACHE_EXTRAFILES
This is similar to CCACHE_EXTRAFILES and ICECC_EXTRAFILES: the extra
files are hashed for caching and packaged for sccache-dist.

Possibly the same variable CCACHE_EXTRAFILES could be used instead of
a distinct one?

This can be implemented in one place for the C/C++ compilers, but not
also for Rust, which currently doesn't have extra_hash_files, so that
remains TODO.