Prefer using Tokio as the main task runtime (#33)
* Spawn root compilation task in Tokio context

Since Tokio 0.2, spawning the `tokio::process::Command` must be done in
the Tokio context, so make sure to spawn the compilation root task in
the Tokio runtime that's now available in the `SccacheService`.

This fixes a hang in the `sccache_cargo` integration test.

* Mark ClientToolchains::put_toolchain as synchronous

It seems entirely synchronous, as it doesn't hit any await points and
additionally seems to be trigger-happy with mutex locks in inner calls,
so better be safe and prevent the future reader from thinking that this
function is async-safe.

* Prepare cache stuff to use Tokio as its I/O-bound task pool

* Prepare to prune ThreadPool entirely in favor of Tokio

The reason for that being not to mix the Tokio runtime and the `futures`
executor. While it's reasonable and possible to do so, I believe there's
bigger benefit for now to stick to a single executor.
It would be great if we could keep the code using 'vanilla' `futures`
crate but Tokio support is so baked in (e.g. child process harness) that
it seems that it'd require a lot more change for questionable benefit in
this case.
Doing so may yield another benefit - currently, there seems to be a lot
of disk I/O being done on a dedicated, blocking thread pool. It's possible,
now that we use a single executor, to use Tokio facilities for async I/O
if this proves to be faster when supported properly by the native
environment at some point (e.g. with `io_uring` on Linux).

* Prefer using Tokio runtime for dist. HTTP client

* Prune futures::executor::ThreadPool

* Prefer using Tokio timer again

* Bring expected test output in line with upstream repository

This lets test pass locally on Ubuntu 20.10 with:
gcc (Ubuntu 10.2.0-13ubuntu1) 10.2.0
Ubuntu clang version 11.0.0-2

* Cherry-pick of CI: switch to staging image before CI is green; fix versions output

https://github.com/paritytech/sccache/commit/85286f9e247a9a4c57aa06355be39d67338e4450

Co-authored-by: Denis P <denis.pisarev@parity.io>