Drop Sync bound from the CompilerProxy
The only reason that's there is because the compiler is too conservative
and treats borrows as alive longer than they actually are (even with
explicit `drop`). Work around that by creating an owned future instead,
returned from the `resolve_proxied_executable`.

It's worth noting that `async_trait` returns a `Pin<Box>`ed future with
the 'async lifetime, where we have &'async self as the method receiver.
Because of this, we manually construct the future ourselves with the
'static lifetime.