Simplify Clippy lint setup
The `unknown_clippy_lints` is deprecated and has been integrated into
the rustc `unknown_lints` lint, see
https://rust-lang.github.io/rust-clippy/master/#unknown_clippy_lints.

For the other two lints that are currently allowed, this moves it to the
source code so it's easier to see what's expected to work and be warned
against and what's not, without having to look at how exactly the
current CI is set up. It's common to simply run `cargo clippy` when
working on the project (similarly to `cargo fmt`) so this optimizes for
that scenario.

Some other repos that use this:
- https://github.com/rust-lang/cargo/blob/master/src/cargo/lib.rs
- https://github.com/rust-lang/rls/blob/master/rls/src/lib.rs
- https://github.com/serde-rs/json/blob/master/src/lib.rs