Don't inherit sccache-related env in integration tests
Because the configuration is merged from both the environment and the
configuration file, it's possible to forget about overriding variables
related to one of the backends (e.g. by setting `SCCACHE_REDIS`). To
account for that and not have to explicitly list/remember all of the
supported env vars, we just don't inherit the sccache-related
environment at all when running this test.