ci: Perform nightly CI/CD runs to catch possible breakage
This helps catching issues unrelated to PRs (which are the most frequent
reason to perform CI runs until now) only popping up when these are
opened. Instead, perform sanity nightly runs to catch the breakage beforehand.
A nice side-effect is that this also builds nightly binaries as part of
the workflows.