Add test for cargo env_deps
The previous commit added support for parsing "env_dep"
information from dep_info files. This commit adds a test
for changing an environment variable, that is referenced
in rust code via env!, and asserts that sccache rebuilds
and the rust code uses the new value of the changed
environment variable.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>