Use clang -fminimize-whitespace to increase cache hits. (#1162)
* Add -fminimize-whitespace preprocessor flag when Clang >= 14 to increase cache hits

This proof of concept tries to take advantage of the new `-fminimize-whitespace`
preprocessor option in Clang. `-fminimize-whitespace` was added to Clang with the
intention to help increase the chance of a cache hit in tools like `sccache`
when only whitespace changes.

I also added a `version()` method to CCompilerImpl trait to expose the compiler
version to the preprocessor. This could be useful if version-dependent
features are added to `sccache` in the future.

Feedback on this is definitely welcome!

* Add tests for clang whitespace normalization

* Use clang -fminimize-whitespace to incread cache hits.

* Use ubuntu-22.04 for testing Clang 14

* Move compiler version from CCompiler to each CCompilerImpl.

* Use semver

Co-authored-by: Max Fan <root@max.fan>
Co-authored-by: Michael Kruse <sccache@meinersbur.de>