Increase the MSRV and bump dependencies (#1137)
* Bump the MSRV to 1.58

Bumps the official MSRV to 1.58.
This is likely more than required the upcoming version bumps.

* Upgrade cargo zstd to 0.10

* Bump serial_test to v0.6

Changelog does not mention any breaking changes.

* Bump blake to v1

* Bump rouille to 3.5

Required to bump syslog to v6, since otherwise no common version
can be selected for `time`

* Bump syslog to v6

No changelog available. Depends on rouille bump to 3.5, to get a
common version of ´time´.

* Upgrade JWT to v8

The explicit insecure decode method was removed and replaced by
an option for the Validation struct.
The `exp` field had to be added to JobJwt, since otherwise the validation would fail at runtime.
Apparently the `exp` field is required now, even if the valdiation of `exp` is turned off.