clang depfile arguments CanBeSeperated (#1205)
This is a fix for gcc/clang: Improper handling of some compilation flags that can be spaced, such as -MF #1197

The arguments `MF`, `MT`, and `MQ`, are now `CanBeSeparated`, which enables them to be correctly parsed by sccache. This is because for clang (but not for gcc), these arguments are allowed to not be separated.

I've added tests for the various forms that the arguments could be presented in, which is potentially testing some things out of scope of the actual change, but it was easy to be more through here.