Add support for clang 'no-opaque-pointers' flag
LLVM is transitioning to _opaque pointers_ but in the meanwhile provides
a way to allow old-style pointers usage: `-Xclang -no-opaque-pointers`
flag.

To support caching in projects that make use of this feature `sccache`
needs to know about this flag. Otherwise all compiler invocations using
it result in the *Can't handle UnknownFlag arguments with -Xclang* and
the output objects are not cached.