Add SCCACHE_S3_RW_MODE to enable read-only S3 cache
This adds a similar capability to the existing `SCCACHE_GCS_RW_MODE`
(and most of the code is a simple copy paste of that implementation).
The main caveat is that this defaults to `READ_WRITE` rather than
`READ_ONLY` as in the GCS switch in order to keep the default behavior
of sccache the same.