Fix duration attributed to cache writes in the log
As it is now, what ends up logged as being the time for cache writes is
the compilation duration exclusively, which is also logged on its own.

For extra granularity, we also log the time spent creating the cache
artifact.

While here, also log the compilation durations in cases where the result
is not cached.