Determine support for -fminimize-whitespace via test execution.
Commit da2934f/PR1162 added the -fminimize-whitespace flag when
proprocessing and derived compiler support from the compiler version
(clang >= 14). It unfortunately had a bug in the detection mechanism and
also the unittest such that the flag was never used:

 1. Clang expands the __VERSION__ macro to a string (e.g. "LLVM Clang 14.0.0").
    The additional quotes caused the version check to fail. Originally,
    the --version output was passed to the version check function which
    does not contain such quotes.

 2. The regression test checked for whether the compiler is "clang", but
    not "clang++".

 3. Wrong regex group names used for checking for Apple caused the check
    to always expect a cache miss.

This patch changes the check whether -fnormalize-whitespace is supported
to a test invocation with that flag. Only the regression test continues
to be based on the know versions that support the flag to be independent
on the feature test. When Apple clang one day adds support for the flag,
the test will need to be updated.