snap: drop the daemon
There's not really a point running a system-wide daemon of sccache. At
least not by default. `sccache` forks a local server if one isn't
running already, and does it as the current user. The server needs write
access to the compilation path, so a central server… doesn't work very
well.