nvcc: Handle compiler detection when gcc isn't on the path
To properly handle compiler detction sccache invokes the compiler
with -E. To generate the preprocessed output `nvcc` invokes the host
comiler, which either requires gcc to be on the path or `-ccbin` argument
to be provided.

This PR allows `sccache` to be used with `nvcc` when `gcc` isn't
on the path but the host compiler to be used is included via `-ccbin`.

This was found when building small build images which only had
a compiler with a name like `gcc-aarch64`.