Use rustls for tls & trust-dns-resolver for dns resolution
By default, reqwest uses native-tls, which is openssl on linux and that
makes cross-compilation very hard as they need to compile openssl
themselves and exposes it via pkg-config.

Enabling vendored-openssl does not help there because the dep:openssl is
only enabled by dist-server.

reqwest by default uses libc's getaddrinfo, which is a poor solution
that requires a dedicated thread and it is not nearly as reliable as
trust-dns-resolver, not to mention libc implementation, e.g. musl, might
deliberately leave out some features.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>