Support for Type Library Headers for MSVC
Introduce caching support for MSVC type library headers.
Type library headers are optional extra outputs that appear when the `#import` statement is used.
This fix takes the easy route of just caching these files if they happen to be produced.

The hard (but more correct way) would be to parse the preprocessed C++ and determine if `#import` is ever used, which would enable us to include that information in the cache key and invalidate any previous cache entries that don't contain type library headers but should.
However, parsing C++ is pretty complex and clearing your cache if you get a missing tlh prior to this change is pretty easy.

This fixes #1769