Produce JSON depsfiles from MSVC
MSVC deps files are emitted as JSON, in a format described by the compiler documentation: https://learn.microsoft.com/en-us/cpp/build/reference/sourcedependencies?view=msvc-170

Prior to this commit, sccache will output a clang or gcc style deps file, with dependencies on lines instead of JSON.

After this commit, we remove the code responsible for writing the deps file, and depend on the command we execute on `cl.exe` for preprocessing to write the deps file.

The advantage of this is that we get improved compatibility with tools that use the JSON output deps files from MSVC, such as Unreal Engine's build system.

This fixes #1771