sccache fails to build on non x86_64/aarch64 arches (#1866)
On all arches except x86_64/aarch64 sccache fails to build. This is
due to the use of rustls/ring as a feature in trust-dns-resolver.

However, trust dns resolver isn't actually needed when you set
--no-default-features, since this appears to only be to coerce
features in trust-dns-resolver when reqwest is present.

As a result, trust-dns-resolver should be optional and only added
when required.