sccache now tracks compiler hits/misses on a per language basis.
Previously sccache was actually tracking hits and misses on a
compiler id basis, and guessing on the language.

This is an issue for compilers like `clang` which support C, C++,
and CUDA. In that case we want to clearly state what language
we are getting hit or misses on to help user diagnose issues.