Allow using preprocessor mode from env without a disk config
If we set `SCCACHE_DIRECT=true` but not `disk_dir` or `disk_sz`, we
end up with `None` which sets `use_preprocessor_cache_mode` to `false`.

But if we run:
```
./target/debug/sccache --start-server
./target/debug/sccache --show-stats
```

We can see that it is uses the local disk by default.

After this change, this works:
`SCCACHE_DIRECT=true ./target/debug/sccache  --start-server
./target/debug/sccache --show-stats`