Activate preprocessor cache mode by default
Preprocessor cache mode (or "direct mode" in ccache speak), offers a
measurable improvement for C/C++ caching.

A non-scientific test of compiling Firefox on a Ryzen 7950x with
hot caches saw this mode changing the wall time from ~48s to ~13s.
This is the overall time spent in `mach build` with the only work
needed is to recompile all C++.

As noted in the docs, this applies only to local storage.