Add randomize_readdir test utility
This is a simple libc shim that randomizes the results of readdir().
For some functionality, cache hashes are dependent on the order that
values are returned from fs::read_dir(), which calls lib readdir().
This library can be inserted into the target application's runtime
using LD_PRELOAD=target/debug/librandmize_readdir.so, after which the
results of readdir() will return in random order. It is only intended
to be used while testing sccache.