Don't cache dep file (#2322)
sccache has two modes:
- normal mode
- preprocessor cache mode (aka direct mode in ccache)

In the former case, the preprocessor is invoked and creates the dep file
itself. Whatever sccache does after that is going to overwrite a fresh
file with potentially wrong information, since the cache key is not
indexed on the elements that would affect the contents of the dep file,
resulting in issue #2321.

In the latter case, sccache currently doesn't handle the situation
appropriately (issue #2194), which is why preprocessor cache mode is
automatically disabled when the dependency flags are on the command line
(#2195). Which also means we fallback to the case above.