Detect __has_include as a blocker of direct mode
The documentation of direct mode says that the result will be wrong
if a file has not been included, but would have been if it was
found. It can still happen in some other ways, I guess, but this
one is easy to detect and block direct mode for the current
compilation unit. The detection is very conservative as currently
implemented.