Clarify documentation about "the hash" (aka cache key)
Especially what it means to hash but not cache a header, which
really makes no sense to someone without much knowledge of
implementation details. I think it's related to a cache entry
being for one source file, but containing several candidate
results distinguished by details about the headers they include.
AFAIU, there is no reason to care about that as a user because
it's always digest of various inputs (cache key) in, result or no
result out.
I think the new wording catches the gist of it.

Reference: if system && config.skip_system_headers in
remember_include_file() in c.rs and a bunch of related code.