Add server name info to stderr of remote jobs that ran, but failed
From experience, flaky hardware or transient software errors(*) in
compile clusters are annoying enough to deal with even with the
information on which node a build failed. Trying to find the bad
node by trial and error, even with binary search, is a minor
nightmare and very time-consuming.

(*) This just happened: the result of unpacking a toolchain tarball
was somehow garbled on one node, so it always failed builds with an
error that could plausibly have indicated a build system problem.

Tampering with stderr in the "job ran, but failed" case is probably
OK because:
- Failing job results aren't cached (or used in build outputs)
- It is very uncommon for a build system to have very detailed
  expectations about stderr of failed compile jobs